everything has to be double!

+++ cameras/Cam_t.cpp
@@ -44,7 +44,7 @@ void Cam_t::raytrace(Scene_t* scene) {
     for (unsigned int j = 0; j < image_->size_y_; j++){
         for (unsigned int i = 0; i < image_->size_x_; i++){
             Vec3f col = Vec3f(); // Or declare above?
-            Vec3f pix_vec = Vec3f(1, PI/2.0 + (j - (double)image_->size_y_/2.0 + 0.5)*pixel_span_y, (i - (double)image_->size_x_/2.0 + 0.5)*pixel_span_x); // Is shit after this line
+            Vec3f pix_vec = Vec3f(1, PI/2.0 + ((double)j - (double)image_->size_y_/2.0 + 0.5)*pixel_span_y, ((double)i - (double)image_->size_x_/2.0 + 0.5)*pixel_span_x); // Is shit after this line
             //std::cout << pix_vec[0] << " " << pix_vec[1] << " " << pix_vec[2] << std::endl; // REMOVE
 
             for (unsigned int k = 0; k < subpix_[0]; k++){
@@ -53,7 +53,7 @@ void Cam_t::raytrace(Scene_t* scene) {
                     double jitter_y = unif_(my_rand::rng);
                     double jitter_x = unif_(my_rand::rng);
 
-                    pix_vec += Vec3f(0, (k - (double)subpix_[0]/2 + jitter_y)*subpix_span_y, (l - (double)subpix_[1]/2 + jitter_x)*subpix_span_x); // Is shit after this line
+                    pix_vec += Vec3f(0, ((double)k - (double)subpix_[0]/2 + jitter_y)*subpix_span_y, ((double)l - (double)subpix_[1]/2 + jitter_x)*subpix_span_x); // Is shit after this line
                     //std::cout << pix_vec[0] << " " << pix_vec[1] << " " << pix_vec[2] << std::endl; // REMOVE
                     pix_vec = to_xyz_offset(pix_vec, direction_, horizontal, vertical); // Is shit after this line
